refactor: cleanup .dot files and global types, add vrt flag to storybook
