--- .storybook-docs/tsconfig.json
@@ -1,28 +1,0 @@
-{
-  //   "compilerOptions": {
-  //     "outDir": "build/lib",
-  //     "module": "commonjs",
-  //     "target": "es5",
-  //     "lib": ["es5", "es6", "es7", "es2017", "dom"],
-  //     "sourceMap": true,
-  //     "allowJs": false,
-  //     "jsx": "react",
-  //     "moduleResolution": "node",
-  //     "rootDirs": ["src", "docs"],
-  //     "baseUrl": "src",
-  //     "forceConsistentCasingInFileNames": true,
-  //     "noImplicitReturns": true,
-  //     "noImplicitThis": true,
-  //     "noImplicitAny": true,
-  //     "strictNullChecks": true,
-  //     "suppressImplicitAnyIndexErrors": true,
-  //     "noUnusedLocals": true,
-  //     "declaration": true,
-  //     "allowSyntheticDefaultImports": true,
-  //     "experimentalDecorators": true,
-  //     "emitDecoratorMetadata": true
-  //   },
-  "extends": "../tsconfig",
-  "include": ["../src/**/*", "../stories/**/*", "./**/*", "../.storybook-docs/config.ts"],
-  "exclude": ["../**/*.test.*"]
-}
