+++ integration/tests/all.test.ts
@@ -21,7 +21,7 @@ import { getStorybookInfo } from '../helpers';
 import { common } from '../page_objects';
 
 // mock required for importing trick, otherwise .scss files will throw an error
-jest.mock('../../.storybook/theme_service.ts', () => ({
+jest.mock('../../storybook/theme_service.ts', () => ({
   switchTheme: () => undefined,
 }));
