+++ integration/tests/area_stories.test.ts
@@ -89,7 +89,6 @@ describe('Area series stories', () => {
 
     it('shows only positive values when hiding negative one', async () => {
       const action = async () => {
-        await common.disableAnimations();
         await page.click('.echLegendItem:nth-child(2) .echLegendItem__label');
       };
       await common.expectChartAtUrlToMatchScreenshot(
@@ -100,7 +99,6 @@ describe('Area series stories', () => {
 
     it('shows only negative values when hiding positive one', async () => {
       const action = async () => {
-        await common.disableAnimations();
         await page.click('.echLegendItem:nth-child(1) .echLegendItem__label');
       };
       await common.expectChartAtUrlToMatchScreenshot(
