+++ storybook-docs/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../tsconfig",
+  "include": ["../src/**/*", "../stories/**/*", "./**/*", "config.ts"],
+  "exclude": ["../**/*.test.*"]
+}
