+++ storybook-docs/webpack.config.js
@@ -19,8 +19,7 @@
 
 const path = require('path');
 
-const CircularDependencyPlugin = require('circular-dependency-plugin');
-const webpack = require('webpack');
+const createCompiler = require('@storybook/addon-docs/mdx-compiler-plugin');
 
 const nonce = 'Pk1rZ1XDlMuYe8ubWV3Lh0BzwrTigJQ=';
 const scssLoaders = [
@@ -37,54 +36,8 @@ const scssLoaders = [
   'sass-loader',
 ];
 
-const MAX_CYCLES = 0;
-let numCyclesDetected = 0;
-
 module.exports = async ({ config }) => {
-  config.plugins.push(new webpack.EnvironmentPlugin({ RNG_SEED: null }));
-  config.plugins.push(
-    new CircularDependencyPlugin({
-      onStart() {
-        numCyclesDetected = 0;
-      },
-      onDetected({ paths, compilation }) {
-        if (!/^node_modules\/.+/.test(paths[0])) {
-          numCyclesDetected++;
-          compilation.warnings.push(new Error(paths.join(' -> ')));
-        }
-      },
-      onEnd({ compilation }) {
-        if (numCyclesDetected > MAX_CYCLES) {
-          compilation.errors.push(
-            new Error(`Detected ${numCyclesDetected} cycles which exceeds configured limit of ${MAX_CYCLES}`),
-          );
-        }
-      },
-    }),
-  );
-
-  config.module.rules.push({
-    test: /\.tsx?$/,
-    loader: 'ts-loader',
-    exclude: /node_modules/,
-    options: {
-      configFile: 'tsconfig.json',
-      transpileOnly: true,
-    },
-  });
-
-  config.module.rules.push({
-    test: /\.tsx?$/,
-    include: [path.resolve(__dirname, '../stories/')],
-    exclude: [path.resolve(__dirname, '../stories/utils')],
-    loaders: [
-      {
-        loader: require.resolve('@storybook/source-loader'),
-        options: { parser: 'typescript' },
-      },
-    ],
-    enforce: 'pre',
-  });
+  // config.plugins.push(new webpack.EnvironmentPlugin({ RNG_SEED: null }));
 
   // Replace default css rules with nonce
   config.module.rules = config.module.rules.filter(({ test }) => !test.test('.css'));
@@ -94,7 +47,7 @@ module.exports = async ({ config }) => {
       {
         loader: 'style-loader',
         options: {
-          attributes: {
+          attrs: {
             nonce,
           },
         },
@@ -108,12 +61,12 @@ module.exports = async ({ config }) => {
 
   config.module.rules.push({
     test: /\.scss$/,
-    include: [path.resolve(__dirname, '../.storybook'), path.resolve(__dirname, '../node_modules/@elastic')],
+    include: [path.resolve(__dirname, '../storybook'), path.resolve(__dirname, '../node_modules/@elastic')],
     use: [
       {
         loader: 'style-loader',
         options: {
-          attributes: {
+          attrs: {
             nonce,
           },
         },
@@ -128,10 +81,9 @@ module.exports = async ({ config }) => {
     resourceQuery: /^\?lazy$/,
     use: [
       {
-        loader: 'style-loader',
+        loader: 'style-loader/useable',
         options: {
-          injectType: 'lazyStyleTag',
-          attributes: {
+          attrs: {
             nonce,
           },
         },
@@ -140,7 +92,46 @@ module.exports = async ({ config }) => {
     ],
   });
 
-  config.resolve.extensions.push('.ts', '.tsx');
+  config.module.rules.push({
+    test: /\.(ts|tsx)$/,
+    use: [
+      {
+        loader: require.resolve('babel-loader'),
+        options: {
+          presets: [['react-app', { flow: false, typescript: true, sourceLoaderOptions: null }]],
+        },
+      },
+      {
+        loader: 'ts-loader',
+        options: {
+          configFile: 'tsconfig.json',
+          transpileOnly: true,
+        },
+      },
+      {
+        loader: require.resolve('react-docgen-typescript-loader'),
+      },
+    ],
+  });
+  config.module.rules.push({
+    test: /\.mdx$/,
+    use: [
+      {
+        loader: 'babel-loader',
+        options: {
+          presets: [['react-app', { flow: false, typescript: true }]],
+        },
+      },
+      {
+        loader: '@mdx-js/loader',
+        options: {
+          compilers: [createCompiler({})],
+        },
+      },
+    ],
+  });
+
+  config.resolve.extensions.push('.ts', '.tsx', '.mdx');
 
   return await config;
 };
