+++ storybook/config.ts
@@ -27,7 +27,11 @@ import { switchTheme } from './theme_service';
 import './style.scss';
 
 switchTheme('light');
-preloadIcons();
+
+if (process.env.STORYBOOK_VRT) {
+  preloadIcons();
+  document.querySelector('html')?.classList.add('disable-animations');
+}
 
 addParameters({
   options: {
