+++ storybook/style.scss
@@ -31,18 +31,22 @@ html {
   z-index: 500;
   box-sizing: border-box;
   background-color: blanchedalmond;
+}
 
-  &.disable-animations {
-    *,
-    *::after,
-    *::before {
-      transition-delay: 0s !important;
-      transition-duration: 0s !important;
-      animation-delay: -0.0001s !important;
-      animation-duration: 0s !important;
-      animation-play-state: paused !important;
-      caret-color: transparent !important;
-    }
+html.disable-animations {
+  *,
+  *::after,
+  *::before {
+    transition-delay: 0s !important;
+    transition-duration: 0s !important;
+    animation-delay: -0.0001s !important;
+    animation-duration: 0s !important;
+    animation-play-state: paused !important;
+    caret-color: transparent !important;
+  }
+
+  .euiSaturation:focus .euiSaturation__indicator {
+    animation: none !important; // fix color picker dot
   }
 }
