+++ storybook/tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "../tsconfig",
+  "include": ["../src/**/*", "../stories/**/*", "./**/*", "../storybook-docs/config.docs.ts"],
+  "exclude": ["../**/*.test.*"]
+}
