+++ storybook/webpack.config.js
@@ -19,7 +19,8 @@
 
 const path = require('path');
 
-const createCompiler = require('@storybook/addon-docs/mdx-compiler-plugin');
+const CircularDependencyPlugin = require('circular-dependency-plugin');
+const webpack = require('webpack');
 
 const nonce = 'Pk1rZ1XDlMuYe8ubWV3Lh0BzwrTigJQ=';
 const scssLoaders = [
@@ -36,8 +37,54 @@ const scssLoaders = [
   'sass-loader',
 ];
 
+const MAX_CYCLES = 0;
+let numCyclesDetected = 0;
+
 module.exports = async ({ config }) => {
-  // config.plugins.push(new webpack.EnvironmentPlugin({ RNG_SEED: null }));
+  config.plugins.push(new webpack.EnvironmentPlugin({ RNG_SEED: null }));
+  config.plugins.push(
+    new CircularDependencyPlugin({
+      onStart() {
+        numCyclesDetected = 0;
+      },
+      onDetected({ paths, compilation }) {
+        if (!/^node_modules\/.+/.test(paths[0])) {
+          numCyclesDetected++;
+          compilation.warnings.push(new Error(paths.join(' -> ')));
+        }
+      },
+      onEnd({ compilation }) {
+        if (numCyclesDetected > MAX_CYCLES) {
+          compilation.errors.push(
+            new Error(`Detected ${numCyclesDetected} cycles which exceeds configured limit of ${MAX_CYCLES}`),
+          );
+        }
+      },
+    }),
+  );
+
+  config.module.rules.push({
+    test: /\.tsx?$/,
+    loader: 'ts-loader',
+    exclude: /node_modules/,
+    options: {
+      configFile: 'tsconfig.json',
+      transpileOnly: true,
+    },
+  });
+
+  config.module.rules.push({
+    test: /\.tsx?$/,
+    include: [path.resolve(__dirname, '../stories/')],
+    exclude: [path.resolve(__dirname, '../stories/utils')],
+    loaders: [
+      {
+        loader: require.resolve('@storybook/source-loader'),
+        options: { parser: 'typescript' },
+      },
+    ],
+    enforce: 'pre',
+  });
 
   // Replace default css rules with nonce
   config.module.rules = config.module.rules.filter(({ test }) => !test.test('.css'));
@@ -47,7 +94,7 @@ module.exports = async ({ config }) => {
       {
         loader: 'style-loader',
         options: {
-          attrs: {
+          attributes: {
             nonce,
           },
         },
@@ -61,12 +108,12 @@ module.exports = async ({ config }) => {
 
   config.module.rules.push({
     test: /\.scss$/,
-    include: [path.resolve(__dirname, '../.storybook'), path.resolve(__dirname, '../node_modules/@elastic')],
+    include: [path.resolve(__dirname, '../storybook'), path.resolve(__dirname, '../node_modules/@elastic')],
     use: [
       {
         loader: 'style-loader',
         options: {
-          attrs: {
+          attributes: {
             nonce,
           },
         },
@@ -81,9 +128,10 @@ module.exports = async ({ config }) => {
     resourceQuery: /^\?lazy$/,
     use: [
       {
-        loader: 'style-loader/useable',
+        loader: 'style-loader',
         options: {
-          attrs: {
+          injectType: 'lazyStyleTag',
+          attributes: {
             nonce,
           },
         },
@@ -92,46 +140,7 @@ module.exports = async ({ config }) => {
     ],
   });
 
-  config.module.rules.push({
-    test: /\.(ts|tsx)$/,
-    use: [
-      {
-        loader: require.resolve('babel-loader'),
-        options: {
-          presets: [['react-app', { flow: false, typescript: true, sourceLoaderOptions: null }]],
-        },
-      },
-      {
-        loader: 'ts-loader',
-        options: {
-          configFile: 'tsconfig.json',
-          transpileOnly: true,
-        },
-      },
-      {
-        loader: require.resolve('react-docgen-typescript-loader'),
-      },
-    ],
-  });
-  config.module.rules.push({
-    test: /\.mdx$/,
-    use: [
-      {
-        loader: 'babel-loader',
-        options: {
-          presets: [['react-app', { flow: false, typescript: true }]],
-        },
-      },
-      {
-        loader: '@mdx-js/loader',
-        options: {
-          compilers: [createCompiler({})],
-        },
-      },
-    ],
-  });
-
-  config.resolve.extensions.push('.ts', '.tsx', '.mdx');
+  config.resolve.extensions.push('.ts', '.tsx');
 
   return await config;
 };
