+++ tsconfig.lint.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "allowJs": true
   },
-  "include": ["./**/*", "./.*.js", "./.playground/**/*", "./.storybook/**/*", "./.storybook-docs/**/*"]
+  "include": ["./**/*", "./.*.js"]
 }
